requirements.txt
fastapi
uvicorn
jinja2
aiofiles
python-multipart
pyobjc-framework-Vision
pyobjc-core
pyobjc

